// packages/twenty-server/src/engine/metadata-modules/ai/ai-agent/types/browsing-context.type.ts
export type BrowsingContextType =
  | {
      type: 'recordPage';
      objectNameSingular: string;
      recordId: string;
      pageLayoutId?: string;
      activeTabId?: string | null;
    }
  | {
      type: 'listView';
      objectNameSingular: string;
      viewId: string;
      viewName: string;
      filterDescriptions: string[];
    }
  | {
      type: 'validationRule';
      objectMetadataId: string;
      objectNameSingular: string;
      validationRuleId?: string;
    };